fix #290 where event are not batched on linger (#306)
**Root cause:**
From awaiting future returned by `kafka_producer.send`, which resolves only after the record is sent, not when it is queued. Which blocks the main loop (`while let Some(Ok(record)) = stream.next().await`)

**Solution**
Use `tokio::spawn` to delegate concurrent-asynchronous polling of the `futures` to tokio's built-in runtime which is non-blocking.

**Notes/questions for posterity**
- With the blocking mechanism in the main loop gone and the usage of tokio::spawn, would we need a backpressure mechanism to avoid memory over-allocation?

**Additional notes:**
There is much change in the code because I had to flatten all the code in the main function to make sense of the lifetimes.