test: spawn connectors using kubernetes access instead of `fluvio connector create` (#323)
Fixes https://github.com/infinyon/fluvio-connectors/issues/311

This PR adds `connector-deploy` binary that can be used to create a kubernetes deployment in the current namespace and cluster. This is done in order to remove the usage of `fluvio connector create` and `delete` for CI tests

Example:
```
cargo run --bin connector-deploy  -- apply --config /path/to/connector-config.yaml
```

Also we can use the delete subcommand and the print subcomand  to delete and print the k8 deploment.